Добавляет абстрактный класс, рефакторит вспомогательные модули
