+++ src/view/filters.js
@@ -1,4 +1,4 @@
-import {createElement} from '../util';
+import AbstractView from './abstract';
 
 const createFilterItemTemplate = (filters, isChecked) => {
   const {name, count} = filters;
@@ -28,26 +28,14 @@ const createFiltersTemplate = (filterItems) => {
               </form>`;
 };
 
-export default class Filters {
+export default class Filters extends AbstractView{
   constructor(filters) {
+    super();
     this._filters = filters;
-    this._element = null;
   }
 
   getTemplate() {
     return createFiltersTemplate(this._filters);
-  }
-
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
   }
 }
