+++ src/view/route-info.js
@@ -1,4 +1,4 @@
-import {createElement} from '../util';
+import AbstractView from './abstract';
 
 const createRouteInfoTemplate = () => {
   return `<section class="trip-main__trip-info  trip-info">
@@ -14,24 +14,8 @@ const createRouteInfoTemplate = () => {
           </section>`;
 };
 
-export default class RouteInfo {
-  constructor() {
-    this._element = null;
-  }
-
+export default class RouteInfo extends AbstractView{
   getTemplate() {
     return createRouteInfoTemplate();
-  }
-
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
   }
 }
